test: lift _wait_for and _restamp_cache into the shared test helpers
